pyproject.toml
[tool.poetry]
name = "django-telegram-bot"
version = "0.1.0"
description = "\"Бот для вебинаров\""
authors = ["Хаустов Александр Дмитриевич <ahousetov@gmail.com>"]
readme = "README.md"
packages = [{include = "django_telegram_bot"}]

[tool.poetry.dependencies]
python = "^3.8"
pytz = "^2023.3"
requests = "^2.30.0"
python-dotenv = "^1.0.0"
ipython = "8.5.0"
django = "3.2.9"
django-cors-headers = "3.13.0"
django-debug-toolbar = "3.6.0"
whitenoise = "6.2.0"
gunicorn = "20.1.0"
uvicorn = "0.18.3"
psycopg2-binary = "2.9.3"
dj-database-url = "1.0.0"
celery = "5.2.7"
redis = "4.3.4"
django-celery-beat = "2.3.0"
python-telegram-bot = "13.15"
setuptools = "67.6.1"
black = "^23.3.0"
openpyxl = "^3.1.2"
emoji = "^2.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
